requirements.txt
blinker==1.8.2
cachelib==0.13.0
captcha==0.6.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Session==0.8.0
Flask-Sessionstore==0.4.5
flask_session_captcha==1.4.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
pillow==10.4.0
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
